add UploadImgFile and rename UploadPost component
