+++ src/Components/Upload/UploadPost/index.scss
@@ -1,5 +1,5 @@
 /*
-  Upload Component Styling
+  Upload Post Component Styling
 */
 
 .postUpload {
