+++ src/Pages/HomePage/index.jsx
@@ -5,7 +5,7 @@ import * as PostAction from 'Redux/Actions/PostAction';
 
 import NavBar from 'Components/Main/NavBar';
 import Post from 'Components/Post/Post';
-import Upload from 'Components/Upload';
+import UploadPost from 'Components/Upload/UploadPost';
 
 const defaultProps = {};
 const propTypes = {};
@@ -38,7 +38,7 @@ class HomePage extends Component {
         <NavBar isActive="feed" />
 
         {newPostClicked ? (
-          <Upload
+          <UploadPost
             userId={this.props.user.id}
             userProfileImg={this.props.user.profileImgName}
             userType={this.props.user.type}
